pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-support-ticket-triage"
version = "0.1.0"
description = "Support-ticket triage environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.8.0",
    "requests>=2.32.0",
    "openai>=1.52.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-cov>=4.0.0"]

[project.scripts]
server = "my_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["my_env", "my_env.server"]
package-dir = { "my_env" = ".", "my_env.server" = "server" }
